refactor(api,tab-manager): flatten the one-file platform/auth directory
Same shape as the vocab/opensidian change: platform/auth/auth.ts had a
seam-shaped path with a single impl behind it. The dashboard drops the
last $platform/auth alias for $lib/platform/auth; tab-manager's
relative import loses the auth/auth stutter.
