+++ apps/tab-manager/src/lib/platform/auth.ts
